Store tokens to disk and use for cloning
The repopath->usertoken map is now replaced by files on disk. When a
hook is created, a file is stored in a 'tokens' directory with the name
of the repository (/ is replaced by -: "user-reponame"). The token is
stored in the same way as with the gin-client: UserToken struct is
stored using gob.