Separate cloning code from validation code
A validator function takes a repository path as argument.  This
repository path should be:
- Fully cloned with all annexed content available.
- At the commit required for the validation.

The validator function is responsible for:
- Reading a configuration file (if necessary) for the validator.
- Determining the level of success of the validation (success, error,
warning).
- Rendering the appropriate badge.
- Saving the results to a file.
- Symlinking 'latest' to the newly created results.

Each validation run creates a directory based on the SHA1 of the commit
being validated.  This will help skip subsequent runs of the same
commit.  Since in the current state, the service expects a directory
called 'latest', we symlink 'latest' to the newest commit each time so
it directs to the latest run, but we keep all the previous results
around.