Run validations as asynchronously
- The runValidator() function runs in a goroutine and returns nothing.
- Each run of runValidator() creates its own session session key.  The
key name and description is the commit hash (guaranteed unique).
- The key generation will fail if a second request comes for the same
commit.  If a validation is running for a commit and a new request comes
for the same commit, the new request will fail (which is probably OK).
This will be fixed when we block repeat attempts to run validations on
the same commit.
- The Validate() function, which handles the validation endpoint used by
hooks, returns "200 OK" if the payload is valid, immediately after
starting the async validation function.
- If anything goes wrong in the runValidator() function (or one of the
validators), it calls the writeValFailure() function, which writes the
failure badge to the results directory.