[results.go ...]  Provide page on missing validation results
Closes #73

"Currently when a logged in user views the validator listing page for a specific repository and there are no results for a specific validator yet, the "Results" link for this validator will lead to a blank page with only the text "404 Nothing to see here ...".

Provide a full page including header and footer, an appropriate
information text and a link back to the results overview page of the
repository."

The page provides two links back to the original page as well
as to one-time validation page.