[Dockerfile] Ensure validators are working on image build
Closes #82

"Currently the validators are just installed but not run on docker
build. It might be a good idea to run them once on building the docker
image to ensure that they are working as they are used by the go binary.

As an example: the usage of the nixio changed, the install succeeded,
but the validation did no longer work as intended."